% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cc_list.R
\name{annotate_cc_list}
\alias{annotate_cc_list}
\title{Call AnnotationGenerator on a list of data}
\usage{
annotate_cc_list(data, data_level, data_types, ...)
}
\arguments{
\item{data}{Named list of data.}

\item{data_level}{Character vector of equal length to data, indicating whether each data type is "peptide" or "protein" -level.}

\item{data_types}{Character vector describing types of data stored in data list. e.g. "zscore", "polyclonal".}
}
\description{
Call AnnotationGenerator on a list of data
}
