% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationGenerator.R
\name{AnnotationGenerator}
\alias{AnnotationGenerator}
\title{generate case control annotations for hit data based on library annotation file.}
\usage{
AnnotationGenerator(data, data_level, data_types, annot,
  peptide_col_id_match, protein_col_id_match, peptide_col_id_display,
  protein_col_id_display, description_col_id, flag_col_id,
  pep_aa = "pep_aa")
}
\arguments{
\item{data}{Input data frame.}

\item{data_level}{If data_level is "peptide", then AnnotationGenerator will operate based on the peptide_ parameters. If data_level is "protein", then Annotation_Generator will operate based on the protein_ parameters.}

\item{data_types}{Character vector describing types of data stored in data list. e.g. "zscore", "polyclonal".}

\item{annot}{Library annotation file with named columns.}

\item{peptide_col_id_match}{For data_level=="peptide", the name of a column in annot that corresponds to the first column of data.}

\item{protein_col_id_match}{For data_level=="protein", the name of a column in annot that corresponds to the first column of data.}

\item{protein_col_id_display}{The name of a column in annot that corresponds to identifiers to display in the markdown report corresponding to each protein.}

\item{description_col_id}{The name of a column in annot that corresponds to identifiers to display in the markdown report corresponding to gene product or similar short protein descriptions to display in the markdown report.}

\item{flag_col_id}{Character string or vector of column  name(s) corresponding to annotation flags to indicate in the markdown report (e.g. of known autoantigens or surface proteins).}

\item{pep_aa}{Character string of column name in annot corresponding to peptide amino acid sequence in the markdown report. Needed for peitopefindr.}

\item{peptide_cold_id_display}{For data_level=="peptide", the name of a column in annot that corresponds to identifiers to display in the markdown report corresponding to each peptide.}
}
\description{
generate case control annotations for hit data based on library annotation file.
}
