% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVARDA_RCP.R
\name{AVARDA_RCPGenerator}
\alias{AVARDA_RCPGenerator}
\title{calculate RCP for AVARDA data}
\usage{
AVARDA_RCPGenerator(AVARDA_paths, AVARDA_grep, case_names, ctrl_names)
}
\description{
calculate RCP for AVARDA data
}
