% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rcp_list.R
\name{compute_rcp_list}
\alias{compute_rcp_list}
\title{Calculate relative control percentile statistic for a list of data.}
\usage{
compute_rcp_list(case, ctrl)
}
\arguments{
\item{case}{Case samples for which to compute RCP.}

\item{ctrl}{Control samples to compute RCP against.}
}
\description{
Case and control should be lists of the same length. Each element of the list
should contain a data frame for which to compute rcp for/against. Corresponding
elements of the case and control lists should be data frames of equal number
of rows.
}
