% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats_list.R
\name{compute_stats_list}
\alias{compute_stats_list}
\title{calculate statistics for list of data}
\usage{
compute_stats_list(case, ctrl, case_rcp, ctrl_rcp, hit_thresh, rcp_thresh,
  stat_test = "Fisher", pval_correction = "BH")
}
\arguments{
\item{case}{Case data list.}

\item{ctrl}{Control data list.}

\item{case_rcp}{Case RCP data list.}

\item{ctrl_rcp}{Control RCP data list.}

\item{hit_thresh}{Enrichment hit threshold.}

\item{rcp_thresh}{RCP hit threshold.}

\item{stat_test}{Statistical test. Currently supports "Fisher"}

\item{pval_correction}{Multiple hypothesis pvalue correction method. BH default.}
}
\description{
calculate statistics for list of data
}
