% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_plan_case_control.R
\name{define_plan_case_control}
\alias{define_plan_case_control}
\title{Setup drake plan for case control report targetse.}
\usage{
define_plan_case_control(config_name = "config.tsv")
}
\arguments{
\item{config_name}{Path to configuration spreadsheet.}
}
\description{
Setup drake plan for case control report targetse.
}
