% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hits.R
\name{filter_hits}
\alias{filter_hits}
\title{placeholder}
\usage{
filter_hits(data, min_hits_enrichment, min_hits_rcp, pval_thresh,
  case_over_control = TRUE)
}
\arguments{
\item{data}{Input data frame.}

\item{min_hits_enrichment}{Threshold value for data_stats$Case.Hits.}

\item{min_hits_rcp}{Threshold value for data_stats$RCP.Hits.Case}

\item{pval_thresh}{Threshold Value for PValue.Min}

\item{case_over_control}{Logical describing whether or not to only keep hits
for which hit frequency in cases is higher than hit frequency in controls.}
}
\description{
placeholder
}
