% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hits_list.R
\name{filter_hits_list}
\alias{filter_hits_list}
\title{filter hits for data list.}
\usage{
filter_hits_list(data_stats, min_hits_enrichment, min_hits_rcp,
  pval_thresh, case_over_control = TRUE, bind_rows = FALSE)
}
\arguments{
\item{data_stats}{Input data list.}

\item{min_hits_enrichment}{Threshold value for data_stats$Case.Hits.}

\item{min_hits_rcp}{Threshold value for data_stats$RCP.Hits.Case}

\item{pval_thresh}{Threshold Value for Fisher.PValue.Min}

\item{case_over_control}{Logical describing whether or not to only keep hits
for which hit frequency in cases is higher than hit frequency in controls.}

\item{bind_rows}{Logical determining wiether dplyr::bind_rows is called at the end.}
}
\description{
filter hits for data list.
}
