% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_sample_list.R
\name{gather_sample_list}
\alias{gather_sample_list}
\title{Return a list of subsetinput data based on specified data types, file paths,
and sample names.}
\usage{
gather_sample_list(data_types, input_files, sample_names)
}
\arguments{
\item{data_types}{Character vector, for example "MLXP", "MLXP_promax", "MLXP_polycl"}

\item{input_files}{Paths to data files from which to extract relevant data. Structured as a list, with each element of the list containing a character vector of paths corresponding to data_type of the same index.}

\item{sample_names}{Sample names corresponding to relevant column names in data files.}
}
\value{
Named list. Name of each element is corresponding data_type.
And value of each element is table of corresponding data.
}
\description{
Return a list of subsetinput data based on specified data types, file paths,
and sample names.
}
