% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot1_hitfreq.R
\name{plot1_hitfreq}
\alias{plot1_hitfreq}
\title{Plot hit frequency vs median hit score of cases}
\usage{
plot1_hitfreq(data, data_types)
}
\arguments{
\item{data}{Filtered hits data frame}

\item{data_types}{List of desired facets in order.}
}
\description{
Plot hit frequency vs median hit score of cases
}
