% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2_hitscore.R
\name{plot2_hitscore}
\alias{plot2_hitscore}
\title{second plot}
\usage{
plot2_hitscore(data, data_types)
}
\arguments{
\item{data}{Filtered hits data frame}

\item{data_types}{List of desired facets in order.}
}
\description{
second plot
}
