% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3_pval.R
\name{plot3_pval}
\alias{plot3_pval}
\title{third plot}
\usage{
plot3_pval(data, data_types)
}
\arguments{
\item{data}{Filtered hits data frame}

\item{data_types}{List of desired facets in order.}
}
\description{
third plot
}
