% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_AVARDA_candidate_table.R
\name{prepare_AVARDA_candidate_table}
\alias{prepare_AVARDA_candidate_table}
\title{prepare_AVARDA_candidate_table}
\usage{
prepare_AVARDA_candidate_table(AVARDA_filtered)
}
\description{
prepare_AVARDA_candidate_table
}
