% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_AVARDA_candidate_table.R
\name{prepare_AVARDA_candidate_table_html}
\alias{prepare_AVARDA_candidate_table_html}
\title{prepare_AVARDA_candidate_table_html}
\usage{
prepare_AVARDA_candidate_table_html(data)
}
\description{
prepare_AVARDA_candidate_table_html
}
