% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_AVARDA_clustergram_data.R
\name{prepare_AVARDA_clustergram_data}
\alias{prepare_AVARDA_clustergram_data}
\title{prepare_AVARDA_clustergram_data}
\usage{
prepare_AVARDA_clustergram_data(AVARDA_case_data, AVARDA_filtered,
  pval_thresh = 0.05)
}
\description{
prepare_AVARDA_clustergram_data
}
