% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_candidate_table.R
\name{prepare_candidate_table}
\alias{prepare_candidate_table}
\title{candidate antigen generator script}
\usage{
prepare_candidate_table(data)
}
\arguments{
\item{data}{annotated_data target.}
}
\description{
Further process annotated_data target by collapsing peptide data to protein data in a more reader-digestible format.
}
