% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_candidate_table_html.R
\name{prepare_candidate_table_html}
\alias{prepare_candidate_table_html}
\title{Add kable html stying to candidate table.}
\usage{
prepare_candidate_table_html(data)
}
\arguments{
\item{data}{output from prepare_candidate_table.}
}
\description{
Add kable html stying to candidate table.
}
