% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_clustergram_data.R
\name{prepare_clustergram_data}
\alias{prepare_clustergram_data}
\title{Process hit data for hierarchical clustering.}
\usage{
prepare_clustergram_data(case_rcp, hit_list, binarize_clustergram = TRUE,
  rcp_thresh = 0.95, sample_field_delimiter = ".",
  sample_key_field = 4, AVARDA_RCP = NA)
}
\arguments{
\item{case_rcp}{RCP data from RCPGenerator or compute_rcp_list.}

\item{hit_list}{Subset candidate data from filter_hits or filter_hits_list.}

\item{binarize_clustergram}{Logical controlling whether or not clustergram will be binarized above/below rcp_thresh.}

\item{rcp_thresh}{Threshold value to binarize against if binarize_clustergram == TRUE.}

\item{sample_field_delimiter}{Field delimiter in sample names (column names of case_rcp.)}

\item{sample_key_field}{Which field(s) to retrain from sample names when separated by sample_field delimiter.}

\item{AVARDA_RCP}{Optional AVARDA breadth case RCP data to add to clustergram.}
}
\description{
Process hit data for hierarchical clustering.
}
