% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_epitope_clustergram_data.R
\name{prepare_epitope_clustergram_data}
\alias{prepare_epitope_clustergram_data}
\title{Collapse hit clustergram based on cross-reactive peptide groups defined by epitopefindr.}
\usage{
prepare_epitope_clustergram_data(clustergram,
  epitopeSummary = "data/epitopefindr/epitopeSummary.csv")
}
\arguments{
\item{clustergram}{Sorted data from generate_clustergram}

\item{epitopeSummary}{Path to epitopeSummary output file from epitopefindr.}
}
\description{
Collapse hit clustergram based on cross-reactive peptide groups defined by epitopefindr.
}
