% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AVARDA.R
\name{read_AVARDA}
\alias{read_AVARDA}
\title{read_AVARDA}
\usage{
read_AVARDA(AVARDA_paths, AVARDA_seropos_grep = "full_BH_p_value",
  AVARDA_breadth_grep = "full_filtered_evidence_number")
}
\description{
read_AVARDA
}
