% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_from_template.R
\name{render_from_template}
\alias{render_from_template}
\title{Pass parameters to R markdown case control template and knit html output file.}
\usage{
render_from_template(output_file = paste0(format(Sys.Date(),
  "\%Y\%m\%d"), "_CaseControl_Report.html"), output_dir = getwd())
}
\arguments{
\item{output_file}{Name of generated html report.}

\item{output_dir}{Directory in which to write report.}
}
\description{
Pass parameters to R markdown case control template and knit html output file.
}
