% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsGenerator.R
\name{StatsGenerator}
\alias{StatsGenerator}
\title{performs a first round of
calculations on the data such as hit frequency and Fisher's test. These
calculations will be used in subsequent graphing.}
\usage{
StatsGenerator(case, ctrl, case_rcp, ctrl_rcp, enrichment_thresh = 10,
  rcp_thresh = 0.95, stat_test = "Fisher", pval_correction = "BH",
  note_range = 1)
}
\arguments{
\item{case}{Case data}

\item{ctrl}{Control data}

\item{case_rcp}{Case RCP data}

\item{ctrl_rcp}{Control RCP data}

\item{enrichment_thresh}{Enrichment hit threshold.}

\item{rcp_thresh}{RCP hit threshold.}

\item{stat_test}{Statistical test. Currently supports "Fisher"}

\item{pval_correction}{Multiple hypothesis pvalue correction method. BH default.}
}
\description{
performs a first round of
calculations on the data such as hit frequency and Fisher's test. These
calculations will be used in subsequent graphing.
}
