% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsGenerator_AVARDA.R
\name{StatsGenerator_AVARDA}
\alias{StatsGenerator_AVARDA}
\title{generate statistics from AVARDA output}
\usage{
StatsGenerator_AVARDA(AVARDA_case_data, AVARDA_ctrl_data,
  seropos_pval = 0.05, pval_correction = "BH")
}
\description{
generate statistics from AVARDA output
}
