% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{subset data to only contain rows that match a template.}
\usage{
subset_data(data, template)
}
\arguments{
\item{data}{List of larger data frames to subset.}

\item{template}{list of smaller data frames to subset against.}
}
\description{
subset data to only contain rows that match a template.
}
