% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_case_ctrl.R
\name{config_case_control}
\alias{config_case_control}
\title{setup configurations for this case control study}
\usage{
config_case_control(case_names = c(), ctrl_names = c(),
  input_dirs = c(), data_types = c("enrichment", "promax", "polycl"),
  rcp_thresh = 10, annot_path = c(), epf_params = c())
}
\description{
setup configurations for this case control study
}
