% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_clustergram.R
\name{generate_clustergram}
\alias{generate_clustergram}
\title{generates clustergram of hits}
\usage{
generate_clustergram(clustergram_rawdata, ncolors = 2)
}
\arguments{
\item{clustergram_rawdata}{output from prepare_clustergram_data or prepare_epitope_clustergram_data}

\item{ncolors}{Number of colors to show in clustergram}
}
\description{
generates clustergram of hits
}
