% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCPGenerator.R
\name{RCPGenerator}
\alias{RCPGenerator}
\title{RCPGenerator calculates RCP, Relative Control Percentile, a case-control
statistic describing the percentile from 0.0 to 1.0 of one score against
that of a control population. For example, if a control population consisted
of three scores, 4.1, 5.0, 6.8, and 10.1, then a score of 5.0 would be
assigned an RCP of 0.25, because 5.0 is higher than 25% of the control
population. In this implementation, ties count as losses.}
\usage{
RCPGenerator(case, ctrl, verbose = FALSE)
}
\arguments{
\item{case}{Data frame of cases.}

\item{ctrl}{Data frame of controls.}

\item{verbose}{Logical to print status updates to console.}
}
\description{
RCPGenerator calculates RCP, Relative Control Percentile, a case-control
statistic describing the percentile from 0.0 to 1.0 of one score against
that of a control population. For example, if a control population consisted
of three scores, 4.1, 5.0, 6.8, and 10.1, then a score of 5.0 would be
assigned an RCP of 0.25, because 5.0 is higher than 25% of the control
population. In this implementation, ties count as losses.
}
