% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_from_template.R
\name{render_from_template}
\alias{render_from_template}
\title{Pass parameters to R markdown case control template and knit html output file.}
\usage{
render_from_template(output_file = paste0(format(Sys.Date(),
  "\%Y\%m\%d"), "_CaseControl_Report.html"), output_dir = getwd(),
  set_title, config, candidate_table_flagged, candidate_table_full, graphs,
  cluster1, motifs, cluster2)
}
\arguments{
\item{output_file}{Name of generated html report.}

\item{output_dir}{Directory in which to write report.}

\item{set_title}{Character string to be used for report title.}

\item{config}{Configuration table used in phipcc:define_plan_case_control. After running drake::make on this plan, config can be retreived by calling drake::readd(config).}

\item{candidate_table_flagged}{Kable html object of flagged/annotated candidate hits to display in report. Retrieved by calling drake::readd(candidate_table_flagged_html).}

\item{candidate_table_full}{Kable html object of all candidate hits to display in report. Retrieved by calling drake::readd(candidate_table_html).}

\item{graphs}{List of ggplot objects to be plotted in report. Retreived by calling list(drake::readd(plot1), drake::readd(plot2), drake::readd(plot3)).}

\item{cluster1}{List of gplots::heatmap.2 object and sorted data. Retreived by calling drake::readd(clustergram1).}

\item{motifs}{Directory path to epitopefindr motifs folder. Typically paste0(getwd(), "/data/epitopefindr/intermediate_files/msa/").}

\item{cluster2}{List of gplots::heatma.2 object after collapsing epitopefindr groups. Retreived by calling drake::readd(clustergram2)}
}
\description{
Pass parameters to R markdown case control template and knit html output file.
}
\seealso{
\code{\link{define_plan_case_control}}
}
